+++ app/database/base.py
@@ -3,6 +3,6 @@ from sqlalchemy.ext.declarative import declarative_base
 Base = declarative_base()
 
 
-from app.data_sources.strong.models import *
-from app.data_sources.amazon.kindle.models import *
-from app.data_sources.google.fitbit.models import *
+from app.models.strong import *
+from app.models.kindle import *
+from app.models.fibit import *
