+++ app/models/kindle.py
