+++ app/models/strong.py
